+++ OneModel.Compression/Decompression/TestResult.cs
@@ -10,17 +10,17 @@ namespace OneModel.Compression.Decompression
 
         public string Type { get; set; }
 
-        public int PhysicalSize { get; set; }
+        public long PhysicalSize { get; set; }
 
-        public int HeadersSize { get; set; }
+        public long HeadersSize { get; set; }
 
         public string CodePage { get; set; }
 
         public bool Ok { get; set; }
 
-        public int Size { get; set; }
+        public long Size { get; set; }
 
-        public int CompressedSize { get; set; }
+        public long CompressedSize { get; set; }
 
     }
 }
